Refactor portfolio management components and update routing
- Moved `ListaAtivos` component to the dashboard and updated its import paths in tests.
- Updated routing in `CalculadoraBalanceamento`, `HistoricoSimulacoes`, `LoginPage`, and home page to redirect to the new dashboard.
- Created new components: `GraficosCarteira`, `ListaAtivos`, `MetricasCards`, `ModaisAtivos`, and `ResumoCarteira` for the dashboard.
- Implemented a new layout for the dashboard using `DashboardLayout`.
- Enhanced the dashboard page with state management for modals and stock actions.
- Updated navigation items in `AppShellEnhanced` to reflect the new dashboard structure.
