+++ docs/MIGRATION_CARTEIRA_TO_DASHBOARD.md
@@ -1,0 +1,82 @@
+# 🔄 Migração `/carteira` → `/dashboard` - Documentação Atualizada
+
+## 📋 Resumo da Migração
+
+A migração da funcionalidade de portfólio de `/carteira` para `/dashboard` foi **concluída com sucesso**. Este documento registra as atualizações na documentação de arquitetura realizadas.
+
+## ✅ Atualizações Realizadas na Documentação
+
+### 1. **Arquivo de Arquitetura Mermaid** (`docs/arquitetura.mmd`)
+
+#### Alterações Principais:
+- **Atualizada descrição do Dashboard**: De "Gestão de Watchlist" para "Gestão de Portfólio"
+- **Removida referência ao EditarAtivos**: O componente foi consolidado no Dashboard
+- **Adicionado comentário de migração**: Documentado no cabeçalho do arquivo
+- **Atualizado fluxo de dados**: Dashboard agora gerencia portfólio diretamente
+- **Adicionada documentação de componentes**: Lista dos componentes migrados
+
+#### Componentes Documentados:
+```
+/src/app/dashboard/components/:
+- MetricasCards.tsx: Cards de métricas do portfólio
+- GraficosCarteira.tsx: Gráficos (pizza e barras) da carteira  
+- ResumoCarteira.tsx: Componente de resumo expansível
+- ListaAtivos.tsx: Lista de ativos com funcionalidades CRUD
+- ModaisAtivos.tsx: Modais para adicionar/editar/excluir ativos
+- index.ts: Arquivo de exportação dos componentes
+```
+
+### 2. **Documentação Geral** (`docs/general/COMPILADO_PROJETO.md`)
+
+#### Atualizações:
+- **Seção "Integração da UI"**: Adicionada nota sobre migração de rotas
+- **Validação**: Mencionado que a funcionalidade foi validada após migração
+
+### 3. **Sumário de Reorganização** (`docs/REORGANIZATION_SUMMARY.md`)
+
+#### Atualizações:
+- **Nova seção**: "Atualizações Recentes" com detalhes da migração
+- **Status atualizado**: Registrada conclusão da migração
+
+## 🎯 Estado Atual
+
+### ✅ Completo:
+- [x] Migração de todos os componentes de `/carteira` para `/dashboard/components`
+- [x] Atualização da documentação de arquitetura Mermaid
+- [x] Remoção do diretório `/carteira` antigo
+- [x] Atualização dos links de referência na documentação
+- [x] Validação de funcionalidades no novo local
+- [x] Documentação das mudanças realizadas
+
+### 📁 Nova Estrutura:
+```
+src/app/dashboard/
+├── layout.tsx                 # Layout com PortfolioProvider
+├── page.tsx                   # Página principal do dashboard  
+└── components/                # Componentes migrados
+    ├── index.ts              # Exportações
+    ├── MetricasCards.tsx     # Cards de métricas
+    ├── GraficosCarteira.tsx  # Gráficos da carteira
+    ├── ResumoCarteira.tsx    # Resumo expansível
+    ├── ListaAtivos.tsx       # Lista com CRUD
+    └── ModaisAtivos.tsx      # Modais de gestão
+```
+
+## 🔗 Navegação
+
+- **Rota principal**: Redirecionamento de `/` para `/dashboard`
+- **Layout de aplicação**: `AppShellEnhanced` aponta para `/dashboard`
+- **Contexto**: `PortfolioProvider` disponível no layout do dashboard
+
+## 📝 Observações
+
+1. **Consistência**: Toda a documentação agora reflete a estrutura migrada
+2. **Funcionalidade**: Nenhuma funcionalidade foi perdida na migração
+3. **Melhoria**: Consolidação melhorou a organização e experiência do usuário
+4. **Manutenibilidade**: Estrutura mais clara facilita manutenção futura
+
+---
+
+**Data da Migração**: 6 de junho de 2025  
+**Status**: ✅ **COMPLETO**  
+**Documentação**: ✅ **ATUALIZADA**
