+++ src/app/page.tsx
@@ -2,5 +2,5 @@ import { redirect } from 'next/navigation';
 
 export default function Home() {
   // Redirecionar para a página unificada de carteira
-  redirect('/carteira');
+  redirect('/dashboard');
 }
